December 2020 Release

+++ lib/cybersource_rest_client/models/v1_file_details_get200_response_file_details.rb
@@ -57,7 +57,7 @@ module CyberSource
         :'last_modified_time' => :'DateTime',
         :'date' => :'Date',
         :'mime_type' => :'String',
-        :'size' => :'Integer'
+        :'size' => :'Float'
       }
     end
